requirements.txt
amqp==2.5.2
anyjson==0.3.3
billiard==3.6.1.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cx-Oracle==7.1.3
Django==2.1.7
django-oauth-toolkit==1.2.0
django-rest-swagger==2.2.0
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.8
importlib-metadata==1.1.0
itypes==1.1.0
Jinja2==2.10.1
kombu==4.6.6
MarkupSafe==1.1.1
more-itertools==8.0.0
oauthlib==3.0.1
openapi-codec==1.3.2
psycopg2==2.7.7
psycopg2-binary==2.8.2
pytz==2018.9
PyYAML==5.1
redis==3.3.11
requests==2.21.0
requests-file==1.4.3
simplejson==3.16.0
six==1.12.0
tldextract==2.2.1
uritemplate==3.0.0
urllib3==1.24.2
vine==1.3.0
whitenoise==3.3.1
zipp==0.6.0
